Extract IME handling to ime.rs (refs #68)
Moves update_ime_position() and render_ime_preedit() out of main.rs
into a dedicated ime.rs module. These are thin &self methods that
report the focused cursor as an IME rect and render the preedit
overlay at the cursor position.

First extraction in amux-app phase 4 decomposition. Mirrors wezterm-
gui's per-concern termwindow/ layout (keyevent.rs, mouseevent.rs, etc).

No behavior change — pure mechanical refactor.
